Extract Jackson into a separate module
