+++ ninja-core/src/main/java/ninja/bodyparser/BodyParserEngineManagerImpl.java
@@ -31,30 +31,24 @@ import com.google.inject.Singleton;
 
 @Singleton
 public class BodyParserEngineManagerImpl implements BodyParserEngineManager {
-    
+
     // Keep a reference of providers rather than instances, so body parser engines
     // don't have to be singleton if they don't want
     private final Map<String, Provider<? extends BodyParserEngine>> contentTypeToBodyParserMap;
 
     @Inject
     public BodyParserEngineManagerImpl(Provider<BodyParserEnginePost> bodyParserEnginePost,
-                                       Provider<BodyParserEngineJson> bodyParserEngineJson,
-                                       Provider<BodyParserEngineXml> bodyParserEngineXml,
                                        Injector injector) {
-        
-        
+
+
         Map<String, Provider<? extends BodyParserEngine>> map = Maps.newHashMap();
 
         // First put the built in ones in, this is so they can be overridden by
         // custom bindings
         map.put(bodyParserEnginePost.get().getContentType(),
                 bodyParserEnginePost);
-        map.put(bodyParserEngineJson.get().getContentType(),
-                bodyParserEngineJson);
-        map.put(bodyParserEngineXml.get().getContentType(),
-                bodyParserEngineXml);
-        
-        
+
+
         // Now lookup all explicit bindings, and find the ones that implement
         // BodyParserEngine
         for (Map.Entry<Key<?>, Binding<?>> binding : injector.getBindings()
@@ -68,7 +62,7 @@ public class BodyParserEngineManagerImpl implements BodyParserEngineManager {
         }
 
         contentTypeToBodyParserMap = ImmutableMap.copyOf(map);
-        
+
 
     }
