+++ ninja-core/src/main/java/ninja/template/TemplateEngineManagerImpl.java
@@ -38,21 +38,9 @@ public class TemplateEngineManagerImpl implements TemplateEngineManager {
     private final Map<String, Provider<? extends TemplateEngine>> contentTypeToTemplateEngineMap;
 
     @Inject
-    public TemplateEngineManagerImpl(Provider<TemplateEngineJson> templateEngineJson,
-                                     Provider<TemplateEngineJsonP> templateEngineJsonP,
-                                     Provider<TemplateEngineXml> templateEngineXmlProvider,
-                                     Injector injector) {
+    public TemplateEngineManagerImpl(Injector injector) {
 
         Map<String, Provider<? extends TemplateEngine>> map = new HashMap<String, Provider<? extends TemplateEngine>>();
-
-        // First put the built in ones in, this is so they can be overridden by
-        // custom bindings
-        map.put(templateEngineJson.get().getContentType(),
-                templateEngineJson);
-        map.put(templateEngineJsonP.get().getContentType(),
-                templateEngineJsonP);
-        map.put(templateEngineXmlProvider.get().getContentType(),
-                templateEngineXmlProvider);
 
         // Now lookup all explicit bindings, and find the ones that implement
         // TemplateEngine
