+++ ninja-jackson/src/main/java/ninja/jackson/BodyParserEngineXml.java
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package ninja.bodyparser;
+package ninja.jackson;
 
 import java.io.IOException;
 
 import ninja.ContentTypes;
 import ninja.Context;
+import ninja.bodyparser.BodyParserEngine;
 
 import org.slf4j.Logger;
 
@@ -29,6 +30,7 @@ import com.fasterxml.jackson.dataformat.xml.JacksonXmlModule;
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
+
 import org.slf4j.LoggerFactory;
 
 @Singleton
