+++ ninja-jackson/src/main/java/ninja/utils/XmlMapperProvider.java
