+++ ninja-servlet-archetype-simple/src/main/resources/archetype-resources/pom.xml
@@ -159,6 +159,11 @@
             <artifactId>ninja-freemarker</artifactId>
             <version>${ninja.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.ninjaframework</groupId>
+            <artifactId>ninja-jackson</artifactId>
+            <version>${ninja.version}</version>
+        </dependency>
         <!-- If you want to deploy to a war please -->
         <!-- comment ninja-standalone dependency and  -->
         <!-- uncomment the dependency for ninja-servlet -->
