+++ ninja-servlet-jpa-blog-integration-test/src/main/java/conf/Module.java
@@ -17,6 +17,7 @@
 package conf;
 
 import ninja.freemarker.FreemarkerModule;
+import ninja.jackson.JacksonModule;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Singleton;
@@ -26,10 +27,11 @@ public class Module extends AbstractModule {
     
 
     protected void configure() {
-        
+
         bind(StartupActions.class);
 
         install(new FreemarkerModule());
+        install(new JacksonModule());
     }
 
 }
