+++ pom.xml
@@ -36,6 +36,9 @@
     <modules>
         <module>ninja-core</module>
         <module>ninja-freemarker</module>
+        <module>ninja-jackson</module>
+        <module>ninja-ehcache</module>
+        <module>ninja-memcached</module>
         <module>ninja-servlet</module>
         <module>ninja-standalone</module>
         <module>ninja-test-utilities</module>
@@ -307,6 +310,24 @@
             <dependency>
                 <groupId>org.ninjaframework</groupId>
                 <artifactId>ninja-freemarker</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.ninjaframework</groupId>
+                <artifactId>ninja-jackson</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.ninjaframework</groupId>
+                <artifactId>ninja-ehcache</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.ninjaframework</groupId>
+                <artifactId>ninja-memcached</artifactId>
                 <version>${project.version}</version>
             </dependency>
